Created PhotoRate class.

+++ atomicdataMB/photolossrates.py
@@ -1,24 +1,32 @@
 """Create photoionization database table."""
 import os, os.path
 import glob
+import pandas as pd
+import psycopg2
+import astropy.units as u
 
 
-def make_photo_table(con):
+def make_photo_table(con=None):
     """Create photoionization database table.
 
-    Searches for *.dat in Data/AtomicData/Loss/Photo/
     If multiple reaction rates are found for a reaction, user is prompted
-        to choose the best one.
-    Saves the table as an HTML table for easy viewing. [Not implemented yet
-        because I don't know where to save it]
+    to choose the best one.
+
+    Photoionization and photodissociation reference:
+    Huebner & Mukherjee (2015), Astrophys. Space Sci., 195, 1-294.
     """
-    cur = con.cursor()
+    if con is None:
+        con = psycopg2.connect(host='localhost', database='thesolarsystemmb')
+        con.autocommit = True
 
     # drop the old table if necessary
-    try:
+    cur = con.cursor()
+    cur.execute('select table_name from information_schema.tables')
+    tables = [r[0] for r in cur.fetchall()]
+    if 'photorates' in tables:
         cur.execute('''DROP table photorates''')
-    except:
-        con.rollback()
+    else:
+        pass
 
     # Make the photorates table
     cur.execute('''CREATE TABLE photorates (
@@ -78,3 +86,65 @@ def make_photo_table(con):
                            SET bestvalue=True
                            WHERE reaction=%s''',
                         (r, ))
+
+class PhotoRate:
+    r"""Determine photoreactions and photorates for a species.
+
+    **Parameters**
+
+    species
+        Species to compute rates for.
+
+    aplanet
+        Distance from the Sun. Default is 1 AU. Given as either a numeric
+        type or an astropy quantity with length units.
+
+    **Attributes**
+
+    species
+        Species
+
+    aplanet
+        Distance from the Sun; astropy quanitity with units AU
+
+    rate
+        Reaction rate; astropy quantity with units s^{-1}
+
+    reactions
+        Pandas dataframe with columns for reaction and rate (in s^{-1}) for
+        that reaction
+    """
+    def __init__(self, species, aplanet_=1.*u.AU):
+        with psycopg2.connect(host='localhost',
+                              database='thesolarsystemmb') as con:
+            prates = pd.read_sql(
+                f'''SELECT reaction, kappa
+                    FROM photorates
+                    WHERE species='{species}' and bestvalue=True''', con)
+
+        try:
+            aplanet_.value
+            aplanet = aplanet_
+        except:
+            aplanet = aplanet_*u.AU
+
+        self.species = species
+        self.aplanet = aplanet
+        a0 = 1*u.AU
+
+        # Photo rate adjusted to proper heliocentric distance
+        if len(prates) == 0:
+            print('No photoreactions found')
+            self.reactions = None
+            self.rate = 1e-30/u.s
+        else:
+            prates['kappa'] = prates['kappa'].apply(
+                lambda k: k * (a0/aplanet)**2)
+            self.reactions = prates
+            self.rate = prates['kappa'].sum()/u.s
+
+    def __str__(self):
+        print(f'Species = {self.species}\n'
+              f'Distance = {self.aplanet}\n'
+              f'Rate = {self.rate}')
+        return ''
